+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 1.3.0
+
+* Catch up with redis-doc: d19fb20..6927ef0:
+  * `SET` command support `KEEPTTL` option.
+  * `LPUSHX` accepts multiple elements.
+  * Add commands support for:
+    *  CLUSTER BUMPEPOCH
+    *  CLUSTER FLUSHSLOTS
+    *  CLUSTER MYID
+    *  MODULE LIST
+    *  MODULE LOAD
+    *  MODULE UNLOAD
+    *  PSYNC
+    *  LATENCY DOCTOR
+    *  LATENCY GRAPH
+    *  LATENCY HISTORY
+    *  LATENCY LATEST
+    *  LATENCY RESET
+    *  LATENCY HELP
+
 ## 1.2.0
 
 * Feature: Peek command now displays more friendly, before each "info" will